Extract shared edition create/update transition logic (#860)
Move app.routers.editions and app.mcp.admin.editions's near-identical
create_edition/update_edition bodies into app/services/editions_service.py,
following the pattern already used by rooms_service.py and layouts_service.py:
the deferred active/edition_type application, target-value computation,
implicit exhibitor clearing, deactivate_conflicting_editions call, and
audit-detail assembly now live in one place instead of two copies that could
silently diverge.

editions_service raises HTTPException directly (matching the pre-existing
shared helpers this consolidates) rather than the ServiceError hierarchy;
the REST router calls it unwrapped while the MCP adapter translates
HTTPException to MCPToolError at its own boundary.
